Renames bonebudget app to transactions. Renames bonebudget db to budgetsimple
