+++ api/requirements.txt
@@ -1,6 +1,8 @@
 asgiref==3.6.0
 Django==4.2
 django-cors-headers==4.1.0
+djangorestframework==3.14.0
 psycopg==3.1.9
+pytz==2023.3
 sqlparse==0.4.4
 typing_extensions==4.6.3
