+++ api/transactions/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class BonebudgetConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'bonebudget'
+    name = 'transactions'
